test_setup/requirements.txt
asgiref==3.11.0
Django==5.2
django-dotenv==1.4.2
gunicorn==23.0.0
packaging==25.0
sqlparse==0.5.5
